moved `SettingsLinkVC` to its own file
